@@ -1,6 +1,6 @@
 {
   "name": "oz_ensemble",
-  "version": "1.5.2",
+  "version": "1.6.0",
   "private": true,
   "scripts": {
     "android": "react-native run-android",
@@ -23,27 +23,27 @@
     "@react-native-community/checkbox": "^0.5.8",
     "@react-native-community/datetimepicker": "^2.2.2",
     "@react-native-community/masked-view": "^0.1.11",
-    "@react-native-community/netinfo": "^5.6.1",
+    "@react-native-community/netinfo": "^6.0.0",
     "@react-native-community/push-notification-ios": "^1.8.0",
     "@react-navigation/bottom-tabs": "^5.11.11",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
-    "@sentry/react-native": "^3.2.2",
-    "base-64": "^0.1.0",
+    "@sentry/react-native": "^3.4.2",
+    "base-64": "^1.0.0",
     "dayjs": "^1.11.2",
     "prop-types": "^15.7.2",
     "react": "17.0.2",
-    "react-native": "0.65.1",
-    "react-native-bootsplash": "^2.2.5",
+    "react-native": "0.68.2",
+    "react-native-bootsplash": "^4.1.5",
     "react-native-calendars": "^1.1284.0",
     "react-native-camera": "^3.30.0",
-    "react-native-config": "^1.4.3",
-    "react-native-device-info": "^5.5.3",
+    "react-native-config": "^1.4.5",
+    "react-native-device-info": "^8.4.8",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-mmkv": "^2.4.1",
     "react-native-permissions": "^3.0.5",
-    "react-native-push-notification": "^7.4.0",
-    "react-native-reanimated": "2.3.0-alpha.2",
+    "react-native-push-notification": "^8.1.1",
+    "react-native-reanimated": "2.3.0-beta.2",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-safe-area-view": "^1.1.1",
     "react-native-screens": "^3.4.0",
@@ -52,12 +52,7 @@
     "react-native-swipe-list-view": "^3.1.3",
     "react-native-swiper": "^1.6.0-nightly.5",
     "react-native-webview": "^10.2.3",
-    "react-redux": "^7.1.3",
     "recoil": "^0.7.3-alpha.2",
-    "recompose": "^0.30.0",
-    "redux": "^4.0.5",
-    "redux-persist": "^6.0.0",
-    "reselect": "^4.0.0",
     "styled-components": "^5.0.0",
     "uuid": "^3.4.0",
     "validator": "^12.2.0"
@@ -66,17 +61,13 @@
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
-    "@types/react-native": "^0.67.6",
-    "@types/styled-components": "^5.1.25",
-    "@types/styled-components-react-native": "^5.1.3",
     "babel-jest": "^26.6.3",
-    "eslint": "7.14.0",
+    "eslint": "^7.32.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-native": "^3.11.0",
     "jest": "^26.6.3",
-    "metro-react-native-babel-preset": "^0.66.0",
-    "react-native-codegen": "^0.0.7",
+    "metro-react-native-babel-preset": "^0.67.0",
     "react-test-renderer": "17.0.2"
   },
   "jest": {
