@@ -1,32 +1,81 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import { Alert } from 'react-native';
-import { connect } from 'react-redux';
-import { compose } from 'recompose';
+import { selector, useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 import ButtonPrimary from '../../components/ButtonPrimary';
 import H2 from '../../components/H2';
 import TextStyled from '../../components/TextStyled';
 import UnderlinedButton from '../../components/UnderlinedButton';
 import { TIPIMAIL_API_KEY, TIPIMAIL_API_USER, TIPIMAIL_EMAIL_FROM } from '../../config';
-import { withToast } from '../../services/toast';
+import { consolidatedCatalogSelector, drinksState } from '../../recoil/consos';
+import { useToast } from '../../services/toast';
 import { defaultPadding } from '../../styles/theme';
-import { getHTMLExport } from '../ConsoFollowUp/consoDuck';
+import { getDisplayName, mapDrinkToDose } from '../ConsoFollowUp/drinksCatalog';
+
+export const HTMLExportSelector = selector({
+  key: 'HTMLExportSelector',
+  get: ({ get }) => {
+    const consolidatedCatalog = get(consolidatedCatalogSelector);
+    const drinks = get(drinksState);
+    return formatHtmlTable(drinks, consolidatedCatalog);
+  },
+});
+
+const formatHtmlTable = (drinks, catalog) => {
+  return `
+    <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "https://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+    <html xmlns="https://www.w3.org/1999/xhtml">
+    <head>
+      <title>Bilan des consommations</title>
+      <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+      <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+      <meta name="viewport" content="width=device-width, initial-scale=1.0 " />
+      <style>
+      table { background-color: white; width: 100%; }
+      td, th { border: 1px solid #ddd; text-align: center; }
+      span { font-weight: bold; }
+      </style>
+    </head>
+    <body>
+      <span>Bilan des consommations</span>
+      <table class="table">
+        <tbody>
+          <tr>
+            <th>Date</th>
+            <th>Quantité</th>
+          </tr>
+          ${[...drinks]
+            .sort((drink1, drink2) => {
+              if (drink1.timestamp < drink2.timestamp) return -1;
+              return 1;
+            })
+            .map((drink) => {
+              const doses = mapDrinkToDose(drink, catalog);
+              const name = getDisplayName(drink.drinkKey, drink.quantity, catalog);
+              const time = new Date(drink.timestamp).getLocaleDateAndTime('fr');
+              return `<tr>
+                  <td>${time}</td>
+                  <td>${drink.quantity} ${name} (${doses} dose${doses > 1 ? 's' : ''})</td>
+                </tr>`;
+            })
+            .join('')}
+        </tbody>
+      </table>
+    </body>
+  </html>
+  `;
+};
 
 const emailFormat = (email) => /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,6}$/i.test(email);
-class Export extends Component {
-  state = {
-    email: '',
-  };
-
-  setEmail = (email) => this.setState({ email });
-
-  exportData = async () => {
-    const { email } = this.state;
+const Export = ({ navigation, onBackPress }) => {
+  const [email, setEmail] = useState('');
+  const toast = useToast();
+  const htmlExport = useRecoilValue(HTMLExportSelector);
+  const exportData = async () => {
     if (!emailFormat(email)) {
       Alert.alert('Adresse email non valide');
       return;
     }
-    const { htmlExport, setToast, onBackPress } = this.props;
     const res = await fetch('https://api.tipimail.com/v1/messages/send', {
       method: 'POST',
       headers: {
@@ -52,41 +101,36 @@ class Export extends Component {
       }),
     }).catch((err) => console.log('sendNPS err', err));
     console.log('email sent', res);
-    setToast(`Email envoyé à ${email}`);
+    toast.show(`Email envoyé à ${email}`);
     onBackPress();
   };
 
-  render() {
-    const { email } = this.state;
-    const { navigation } = this.props;
-    return (
-      <Container>
-        <BackButton content="< Retour" onPress={navigation.goBack} bold />
-        <SubContainer>
-          <Title>
-            <TextStyled color="#4030a5">Exporter mes données</TextStyled>
-          </Title>
-          <TextInputStyled
-            value={email}
-            ref={(r) => (this.emailInput = r)}
-            placeholder="Adresse email"
-            onChangeText={this.setEmail}
-            autoCompleteType="email"
-            autoCorrect={false}
-            keyboardType="email-address"
-            autoCapitalize="none"
-            returnKeyType="go"
-            textContentType="emailAddress"
-            onSubmitEditing={this.sendNPS}
-          />
-          <ButtonsContainer>
-            <ButtonPrimary content="Envoyer" disabled={!email} onPress={this.exportData} />
-          </ButtonsContainer>
-        </SubContainer>
-      </Container>
-    );
-  }
-}
+  return (
+    <Container>
+      <BackButton content="< Retour" onPress={navigation.goBack} bold />
+      <SubContainer>
+        <Title>
+          <TextStyled color="#4030a5">Exporter mes données</TextStyled>
+        </Title>
+        <TextInputStyled
+          value={email}
+          placeholder="Adresse email"
+          onChangeText={setEmail}
+          autoCompleteType="email"
+          autoCorrect={false}
+          keyboardType="email-address"
+          autoCapitalize="none"
+          returnKeyType="go"
+          textContentType="emailAddress"
+          onSubmitEditing={exportData}
+        />
+        <ButtonsContainer>
+          <ButtonPrimary content="Envoyer" disabled={!email} onPress={exportData} />
+        </ButtonsContainer>
+      </SubContainer>
+    </Container>
+  );
+};
 
 const Container = styled.View`
   justify-content: center;
@@ -137,8 +181,4 @@ const BackButton = styled(UnderlinedButton)`
   margin-left: -20px;
 `;
 
-const makeStateToProps = () => (state) => ({
-  htmlExport: getHTMLExport(state),
-});
-
-export default compose(connect(makeStateToProps), withToast)(Export);
+export default Export;
