@@ -5,9 +5,9 @@ import styled from 'styled-components';
 const StyledSvg = styled(Svg)``;
 
 const Calendar = ({ color, size, ...props }) => (
-  <StyledSvg width={size} height={size} viewBox="0 0 31.5 31.5" {...props}>
+  <StyledSvg width={size} height={size} viewBox="0 0 24 24" {...props}>
     <Path
-      d="M4 9H6V11H4V9ZM18 4V18C18 19.1 17.1 20 16 20H2C0.89 20 0 19.1 0 18L0.00999999 4C0.00999999 2.9 0.89 2 2 2H3V0H5V2H13V0H15V2H16C17.1 2 18 2.9 18 4ZM2 6H16V4H2V6ZM16 18V8H2V18H16ZM12 11H14V9H12V11ZM8 11H10V9H8V11Z"
+      d="M7 11H9V13H7V11ZM21 6V20C21 21.1 20.1 22 19 22H5C3.89 22 3 21.1 3 20L3.01 6C3.01 4.9 3.89 4 5 4H6V2H8V4H16V2H18V4H19C20.1 4 21 4.9 21 6ZM5 8H19V6H5V8ZM19 20V10H5V20H19ZM15 13H17V11H15V13ZM11 13H13V11H11V13Z"
       fill={color}
     />
   </StyledSvg>
