@@ -1,13 +1,13 @@
-import React, { useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { useNavigation } from '@react-navigation/native';
-import { connect } from 'react-redux';
+import { findNodeHandle } from 'react-native';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import Background from '../../components/Background';
 import DrinksCategory from '../../components/DrinksCategory';
 import HeaderBackground from '../../components/HeaderBackground';
 import TextStyled from '../../components/TextStyled';
 import { makeSureTimestamp } from '../../helpers/dateHelpers';
 import matomo from '../../services/matomo';
-import { checkIfThereIsDrinks, setModalTimestamp } from './consoDuck';
 import Diagram from './Diagram';
 import DiagramHelpModal from './DiagramHelpModal';
 import { BEER, BEER_HALF, drinksCatalog } from './drinksCatalog';
@@ -21,23 +21,40 @@ import {
   Title,
   TopContainer,
 } from './styles';
+import { drinksState, modalTimestampState } from '../../recoil/consos';
 
 const fakeDrinks = [{ drinkKey: BEER_HALF, quantity: 1 }];
 
-const ConsoFollowUp = ({ showWelcomeMessage, setModalTimestamp }) => {
+const ConsoFollowUp = () => {
+  const showWelcomeMessage = !useRecoilValue(drinksState)?.length;
+  const setModalTimestamp = useSetRecoilState(modalTimestampState);
   const [showHelpModal, setShowHelpModal] = useState(false);
   const [selectedBar, setSelectedBar] = useState({});
   const navigation = useNavigation();
+  const scrollViewRef = useRef(null);
 
   const addDrinksRequest = (timestamp) => {
     setModalTimestamp(timestamp);
     navigation.push('ADD_DRINK');
   };
+  const scrollToInput = (ref) => {
+    if (!ref) return;
+    if (!scrollViewRef.current) return;
+    setTimeout(() => {
+      ref.measureLayout(
+        findNodeHandle(scrollViewRef.current),
+        (x, y) => {
+          scrollViewRef.current.scrollTo({ y: y - 100, animated: true });
+        },
+        (error) => console.log('error scrolling', error)
+      );
+    }, 250);
+  };
 
   return (
     <Background color="#39cec0" withSwiperContainer>
       <HeaderBackground />
-      <ScreenBgStyled>
+      <ScreenBgStyled ref={scrollViewRef}>
         <TopContainer>
           <Title>
             <TextStyled color="#4030a5">Suivez votre consommation en unités d'alcool</TextStyled>
@@ -110,19 +127,11 @@ const ConsoFollowUp = ({ showWelcomeMessage, setModalTimestamp }) => {
             <NoDrinkTodayButton timestamp={Date.now()} content="Je n'ai rien bu aujourd'hui !" />
           )}
         </FeedAddConsoTodayContainer>
-        <Feed hideFeed={showWelcomeMessage} />
+        <Feed hideFeed={showWelcomeMessage} scrollToInput={scrollToInput} />
       </ScreenBgStyled>
       <DiagramHelpModal visible={showHelpModal} onCloseHelp={() => setShowHelpModal(false)} />
     </Background>
   );
 };
 
-const makeStateToProps = () => (state) => ({
-  showWelcomeMessage: !checkIfThereIsDrinks(state),
-});
-
-const dispatchToProps = {
-  setModalTimestamp,
-};
-
-export default connect(makeStateToProps, dispatchToProps)(ConsoFollowUp);
+export default ConsoFollowUp;
