@@ -1,10 +1,10 @@
 import { useNavigation } from '@react-navigation/native';
 import React, { useMemo, useState } from 'react';
-import { Text, TouchableOpacity } from 'react-native';
+import { Text, TouchableOpacity, TouchableWithoutFeedback } from 'react-native';
 import Speedometer from 'react-native-speedometer-chart';
-import { connect } from 'react-redux';
 import styled from 'styled-components';
 import { useRecoilValue } from 'recoil';
+import dayjs from 'dayjs';
 import { screenHeight, screenWidth } from '../../styles/theme';
 import H1 from '../../components/H1';
 import H2 from '../../components/H2';
@@ -19,22 +19,23 @@ import GainsCalendar from './GainsCalendar';
 import CocktailGlass from '../../components/Illustrations/CocktailGlassTriangle';
 import Done from '../../components/Illustrations/Done';
 import { drinksCatalog } from '../ConsoFollowUp/drinksCatalog';
-import { daysWithGoalNoDrinkState, maxDrinksPerWeekSelector, previousDrinksPerWeekState } from './recoil';
+import { daysWithGoalNoDrinkState, maxDrinksPerWeekSelector, previousDrinksPerWeekState } from '../../recoil/gains';
 import OnBoardingGain from './OnBoardingGain';
-import { getDaysForFeed, getDailyDoses, getDrinksState } from '../ConsoFollowUp/consoDuck';
-import dayjs from 'dayjs';
+import { dailyDosesSelector, drinksState, feedDaysSelector } from '../../recoil/consos';
 
-const MyGains = ({ days, dailyDoses, drinks }) => {
+const MyGains = () => {
   const navigation = useNavigation();
+  const drinks = useRecoilValue(drinksState);
+  const days = useRecoilValue(feedDaysSelector);
+  const dailyDoses = useRecoilValue(dailyDosesSelector());
   const maxDrinksPerWeekGoal = useRecoilValue(maxDrinksPerWeekSelector);
   const previousDrinksPerWeek = useRecoilValue(previousDrinksPerWeekState);
   const dayNoDrink = useRecoilValue(daysWithGoalNoDrinkState)?.length;
 
   const navigateToGoal = () => {
-    navigation.navigate('GOAL');
-    setShowOnboardingGainModal((show) => !show);
+    navigation.navigate('GAINS_MY_OBJECTIVE');
+    setShowOnboardingGainModal(false);
   };
-  const navigateToEstimation = () => navigation.navigate('ESTIMATION');
   const isOnboarded = useMemo(
     () => !!maxDrinksPerWeekGoal && !!previousDrinksPerWeek.length,
     [maxDrinksPerWeekGoal, previousDrinksPerWeek]
@@ -53,6 +54,7 @@ const MyGains = ({ days, dailyDoses, drinks }) => {
         .length,
     [days, dailyDoses]
   );
+
   const numberDrinkThisWeek = useMemo(
     () =>
       days
@@ -66,8 +68,11 @@ const MyGains = ({ days, dailyDoses, drinks }) => {
   );
 
   const myWeeklyNumberOfDrinksBeforeObjective = useMemo(() => {
-    return previousDrinksPerWeek.reduce((sum, drink) => sum +  drink.quantity*drinksCatalog.find((drinkcatalog)=>drinkcatalog.drinkKey=== drink.drinkKey).doses
-, 0);
+    return previousDrinksPerWeek.reduce(
+      (sum, drink) =>
+        sum + drink.quantity * drinksCatalog.find((drinkcatalog) => drinkcatalog.drinkKey === drink.drinkKey).doses,
+      0
+    );
   }, [previousDrinksPerWeek]);
 
   const myWeeklyExpensesBeforeObjective = useMemo(
@@ -127,7 +132,7 @@ const MyGains = ({ days, dailyDoses, drinks }) => {
           <H1 color="#4030a5">Mes gains</H1>
         </TopTitle>
         {!isOnboarded ? (
-          <TouchableOpacity onPress={() => navigation.navigate('GOAL')}>
+          <TouchableOpacity onPress={() => navigation.navigate('GAINS_MY_OBJECTIVE')}>
             <Description>
               <InfosIcon size={24} />
               <TextDescritpion>
@@ -223,7 +228,7 @@ const MyGains = ({ days, dailyDoses, drinks }) => {
         visible={showOnboardingGainModal}
         hide={() => setShowOnboardingGainModal(false)}
       />
-      <GainsCalendar isOnboarded={isOnboarded} />
+      <GainsCalendar isOnboarded={isOnboarded} setShowOnboardingGainModal={setShowOnboardingGainModal} />
       {!isOnboarded ? (
         <BottomContainer>
           <TopTitle>
@@ -279,22 +284,24 @@ const MyGains = ({ days, dailyDoses, drinks }) => {
             <MyGoalSubContainerInside>
               <PartContainer>
                 <Economy size={20} />
-                <TextStyled> {myWeeklyExpensesBeforeObjective} €</TextStyled>
+                <TextStyled>
+                  {'   '}
+                  {myWeeklyExpensesBeforeObjective} €
+                </TextStyled>
               </PartContainer>
               <PartContainer>
                 <CocktailGlass size={20} />
                 <TextStyled>
-                  {' '}
-                  {myWeeklyNumberOfDrinksBeforeObjective}{' '}
-                  {myWeeklyNumberOfDrinksBeforeObjective > 1 ? 'verres =' : 'verre ='}{' '}
-                  {myWeeklyNumberOfDrinksBeforeObjective}{' '}
-                  {myWeeklyNumberOfDrinksBeforeObjective > 1 ? "doses d'alcool" : "dose d'alcool"}{' '}
+                  {'   '}
+                  {myWeeklyNumberOfDrinksBeforeObjective} verre
+                  {myWeeklyNumberOfDrinksBeforeObjective > 1 ? 's' : ''} ( = {myWeeklyNumberOfDrinksBeforeObjective}{' '}
+                  {myWeeklyNumberOfDrinksBeforeObjective > 1 ? "doses d'alcool" : "dose d'alcool"})
                 </TextStyled>
               </PartContainer>
             </MyGoalSubContainerInside>
           </MyGoalSubContainer>
           <ModifyContainer>
-            <ButtonTouchable onPress={navigateToEstimation}>
+            <ButtonTouchable onPress={() => navigation.navigate('GAINS_ESTIMATE_PREVIOUS_CONSUMPTION')}>
               <TextModify>
                 <TextStyled>Modifier l'estimation</TextStyled>
               </TextModify>
@@ -375,7 +382,7 @@ const Title = styled.View`
 `;
 
 const MyGoalContainer = styled.View`
-  padding: 20px 30px 0px;
+  padding: 20px 30px 100px;
 `;
 
 const MyGoalSubContainer = styled.View`
@@ -405,12 +412,6 @@ const TextModify = styled.Text`
   text-decoration: underline;
 `;
 
-const makeStateToProps = () => (state) => ({
-  drinks: getDrinksState(state),
-  days: getDaysForFeed(state),
-  dailyDoses: getDailyDoses(state),
-});
-
-export default connect(makeStateToProps)(MyGains);
-
 const ButtonTouchable = styled.TouchableOpacity``;
+
+export default MyGains;
